+++ app/models/college.rb
@@ -26,7 +26,7 @@ class College < ApplicationRecord
   validates :dean, presence: true
   validates :subdomain, uniqueness: { case_sensitive: false }
 
-  has_many :users
+  has_many :users, dependent: :nullify
 
   before_validation :set_subdomain
   before_update :freeze_subdomain